refactor: Inline `Taxon#any_nontaxt_references?`

+++ app/helpers/taxon_buttons_helper.rb
@@ -54,8 +54,7 @@ module TaxonButtonsHelper
   end
 
   def delete_unreferenced_taxon_button taxon
-    return if taxon.is_a? Family
-    return if taxon.any_nontaxt_references?
+    return if taxon.is_a?(Family) || Taxa::AnyNonTaxtReferences[taxon]
 
     link_to 'Delete', taxa_path(taxon), method: :delete, class: "btn-warning",
       data: { confirm_with_edit_summary: true }
